A lot of changes again
Download finally working.
Plex could have been misconfigured and lead me to think that there were an issue with mp4 metadatas, but actually it's working... ffmpeg could be useless so, have to check out that.
Filenaming is now done at json creation and so added to json files (could be actually useful for a "auto delete on a specified date" function)
If ffmpeg is still useful, I'll have to be careful about "/" and "\" for filepaths as they actually could have issues depending on the OS (fs in Node is not impacted because it actually check by itself but not the exec command with ffmpeg)
Also added a json folder for failed requests.
Next step, do the wiki, add a "get thumbnail" function and later a web interface to config everything.